rename brewblox-devcon-spark
