+++ tox.ini
@@ -16,7 +16,7 @@ deps =
     -rrequirements-dev.txt
 commands = python -m pytest \
                 --cache-clear \
-                --cov=brewblox_controller_spark \
+                --cov=brewblox_devcon_spark \
                 --cov-branch \
                 --cov-report=term-missing:skip-covered \
                 --cov-fail-under=100 \
